extract time chart
